move euclideanDistance to Util
